refactor: checkIn type
