+++ src/dtos/checkInsDTO.ts
@@ -1,4 +1,11 @@
-import { CheckIn } from '@/types/checkIn1211';
+import { CheckIn, ChildCheckIn } from '@/types/checkIn';
+
+export type ChildCheckInDTO = {
+  email: string;
+  event: string;
+  id: string;
+  user_id: string;
+};
 
 export type CheckInDTO = {
   checkin: {
@@ -11,7 +18,18 @@ export type CheckInDTO = {
   lastname: string;
 };
 
-export const convertCheckInDTOtoCheckIn = (checkinDTO: CheckInDTO): CheckIn => {
+export const ChildCheckInParser = (
+  checkinDTO: ChildCheckInDTO
+): ChildCheckIn => {
+  return {
+    email: checkinDTO.email,
+    event: checkinDTO.event,
+    id: checkinDTO.id,
+    userId: checkinDTO.user_id,
+  };
+};
+
+export const CheckInParser = (checkinDTO: CheckInDTO): CheckIn => {
   return {
     checkIn: {
       email: checkinDTO.checkin.email,
