+++ src/types/checkIn.ts
@@ -1,6 +1,17 @@
-export type CheckIn = {
+export type ChildCheckIn = {
   email: string;
   event: string;
   id: string;
   userId: string;
+};
+
+export type CheckIn = {
+  checkIn: {
+    email: string;
+    event: string;
+    id: string;
+    userId: string;
+  };
+  firstName: string;
+  lastName: string;
 };
