--- src/types/checkIn1211.ts
@@ -1,10 +1,0 @@
-export type CheckIn = {
-  checkIn: {
-    email: string;
-    event: string;
-    id: string;
-    userId: string;
-  };
-  firstName: string;
-  lastName: string;
-};
