+++ src/utils/user.ts
@@ -2,7 +2,7 @@ import { User } from '@/types/user';
 import { getAccessToken, getUserId } from '@/utils/auth';
 import { apiClient } from './axios';
 import { AxiosResponse } from 'axios';
-import { convertUserDTOtoUser, UserDTO } from '@/dtos/userDTO';
+import { UserDTO, UserParser } from '@/dtos/userDTO';
 
 export const getUser = async (): Promise<User | null> => {
   const accessToken = await getAccessToken();
@@ -22,9 +22,8 @@ export const getUser = async (): Promise<User | null> => {
       }
     );
 
-    return convertUserDTOtoUser(res.data.user);
+    return UserParser(res.data.user);
   } catch (error) {
-    console.log('error:', error);
     return null;
   }
 };
